# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MOEX Sentinel API
  version: 0.1.0
paths:
...
  /api/brokers:
    get:
      tags:
      - brokers
      summary: View Broker Settings
      operationId: view_broker_settings_api_brokers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerSettingsSchema'
    post:
      tags:
      - brokers
      summary: Create Broker Settings
      operationId: create_broker_settings_api_brokers_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerDraftSchema'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/brokers/{broker_id}:
    put:
      tags:
      - brokers
      summary: Replace Broker Settings
      operationId: replace_broker_settings_api_brokers__broker_id__put
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerDraftSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - brokers
      summary: Delete Broker Settings
      operationId: delete_broker_settings_api_brokers__broker_id__delete
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/brokers/{broker_id}/check:
    post:
      tags:
      - brokers
      summary: Check Broker Connection
      operationId: check_broker_connection_api_brokers__broker_id__check_post
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/moex_sentinel__views__brokers__BrokerConnectionSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/brokers/{broker_id}/instruments/synchronize:
    post:
      tags:
      - instruments
      summary: Synchronize
      operationId: synchronize_api_brokers__broker_id__instruments_synchronize_post
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationResultSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/brokers/{broker_id}/instruments:
    get:
      tags:
      - instruments
      summary: Instruments
      operationId: instruments_api_brokers__broker_id__instruments_get
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Inactive
      - name: selected_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Selected Only
      - name: lot_price_from
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Lot Price From
      - name: lot_price_to
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Lot Price To
      - name: currency
        in: query
        required: false
        schema:
          type: string
          default: RUB
          title: Currency
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogListResponseSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/brokers/{broker_id}/instruments/{instrument_id}/selection:
    put:
      tags:
      - instruments
      summary: Set Selection
      operationId: set_selection_api_brokers__broker_id__instruments__instrument_id__selection_put
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      - name: instrument_id
        in: path
        required: true
        schema:
          type: string
          title: Instrument Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstrumentSelectionSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogInstrumentSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/brokers/{broker_id}/instruments/{instrument_id}:
    get:
      tags:
      - instruments
      summary: Instrument Details
      operationId: instrument_details_api_brokers__broker_id__instruments__instrument_id__get
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      - name: instrument_id
        in: path
        required: true
        schema:
          type: string
          title: Instrument Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentDetailsSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/instruments/{broker_id}/{instrument_id}/trade:
    post:
      tags:
      - trading-automations
      summary: Create Automation
      operationId: create_automation_api_instruments__broker_id___instrument_id__trade_post
      parameters:
      - name: broker_id
        in: path
        required: true
        schema:
          type: string
          title: Broker Id
      - name: instrument_id
        in: path
        required: true
        schema:
          type: string
          title: Instrument Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAutomationSchema'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/trading-automations/{automation_id}:
    get:
      tags:
      - trading-automations
      summary: View Automation
      operationId: view_automation_api_trading_automations__automation_id__get
      parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          type: string
          title: Automation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/trading-automations/{automation_id}/details:
    get:
      tags:
      - trading-automations
      summary: View Automation Details
      operationId: view_automation_details_api_trading_automations__automation_id__details_get
      parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          type: string
          title: Automation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingAutomationDetailsSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/trading-automations/{automation_id}/hold:
    post:
      tags:
      - trading-automations
      summary: Hold Automation
      operationId: hold_automation_api_trading_automations__automation_id__hold_post
      parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          type: string
          title: Automation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/trading-automations/{automation_id}/resume:
    post:
      tags:
      - trading-automations
      summary: Resume Automation
      operationId: resume_automation_api_trading_automations__automation_id__resume_post
      parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          type: string
          title: Automation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/trading-automations/{automation_id}/close:
    post:
      tags:
      - trading-automations
      summary: Close Automation
      operationId: close_automation_api_trading_automations__automation_id__close_post
      parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          type: string
          title: Automation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AutomationSchema:
      properties:
        id:
          type: string
          title: Id
        broker_id:
          type: string
          title: Broker Id
        account_id:
          type: string
          title: Account Id
        instrument_id:
          type: string
          title: Instrument Id
        state:
          $ref: '#/components/schemas/AutomationState'
        suspended_from_state:
          anyOf:
          - $ref: '#/components/schemas/AutomationState'
          - type: 'null'
        revision:
          type: integer
          title: Revision
        last_sequence_number:
          type: integer
          title: Last Sequence Number
        resume_requested:
          type: boolean
          title: Resume Requested
        strategy:
          $ref: '#/components/schemas/StrategySchema-Output'
        quantity_lots:
          type: integer
          title: Quantity Lots
        average_price:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Average Price
        invested_amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Invested Amount
        realized_pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Realized Pnl
        unrealized_pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Unrealized Pnl
        net_pnl:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Net Pnl
        actual_commissions:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Actual Commissions
        broker_name:
          type: string
          title: Broker Name
        ticker:
          type: string
          title: Ticker
        instrument_name:
          type: string
          title: Instrument Name
      additionalProperties: false
      type: object
      required:
      - id
      - broker_id
      - account_id
      - instrument_id
      - state
      - suspended_from_state
      - revision
      - last_sequence_number
      - resume_requested
      - strategy
      - quantity_lots
      - average_price
      - invested_amount
      - realized_pnl
      - unrealized_pnl
      - net_pnl
      - actual_commissions
      - broker_name
      - ticker
      - instrument_name
      title: AutomationSchema
    AutomationState:
      type: string
      enum:
      - IN_QUEUE
      - OPENING
      - IN_WORK
      - HOLD
      - CLOSED
      - STOPPED
      - COMPLETED
      title: AutomationState
...
    BrokerAdapterSchema:
      properties:
        adapter_code:
          type: string
          title: Adapter Code
        provider_code:
          type: string
          title: Provider Code
        environment_code:
          type: string
          title: Environment Code
        fields:
          items:
            $ref: '#/components/schemas/BrokerFieldDefinitionSchema'
          type: array
          title: Fields
      type: object
      required:
      - adapter_code
      - provider_code
      - environment_code
      - fields
      title: BrokerAdapterSchema
    BrokerDraftSchema:
      properties:
        display_name:
          type: string
          title: Display Name
        provider_code:
          type: string
          title: Provider Code
        environment_code:
          type: string
          title: Environment Code
        adapter_code:
          type: string
          title: Adapter Code
        enabled:
          type: boolean
          title: Enabled
        fields:
          items:
            $ref: '#/components/schemas/BrokerFieldSchema'
          type: array
          title: Fields
        is_test:
          type: boolean
          title: Is Test
        initial_balance:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Initial Balance
      additionalProperties: false
      type: object
      required:
      - display_name
      - provider_code
      - environment_code
      - adapter_code
      - enabled
      - fields
      - is_test
      - initial_balance
      title: BrokerDraftSchema
...
    BrokerFieldSchema:
      properties:
        name:
          type: string
          title: Name
        value:
          type: string
          title: Value
      additionalProperties: false
      type: object
      required:
      - name
      - value
      title: BrokerFieldSchema
    BrokerSchema:
      properties:
        id:
          type: string
          title: Id
        display_name:
          type: string
          title: Display Name
        provider_code:
          type: string
          title: Provider Code
        environment_code:
          type: string
          title: Environment Code
        adapter_code:
          type: string
          title: Adapter Code
        enabled:
          type: boolean
          title: Enabled
        is_test:
          type: boolean
          title: Is Test
        initial_balance:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Initial Balance
        test_account_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Test Account Id
        test_account_funded:
          type: boolean
          title: Test Account Funded
        fields:
          items:
            $ref: '#/components/schemas/BrokerFieldSchema'
          type: array
          title: Fields
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - display_name
      - provider_code
      - environment_code
      - adapter_code
      - enabled
      - is_test
      - initial_balance
      - test_account_id
      - test_account_funded
      - fields
      - created_at
      - updated_at
      title: BrokerSchema
    BrokerSettingsSchema:
      properties:
        adapters:
          items:
            $ref: '#/components/schemas/BrokerAdapterSchema'
          type: array
          title: Adapters
        brokers:
          items:
            $ref: '#/components/schemas/BrokerSchema'
          type: array
          title: Brokers
      type: object
      required:
      - adapters
      - brokers
      title: BrokerSettingsSchema
...
    CatalogInstrumentPriceSchema:
      properties:
        id:
          type: string
          title: Id
        broker_id:
          type: string
          title: Broker Id
        instrument_id:
          type: string
          title: Instrument Id
        figi:
          type: string
          title: Figi
        ticker:
          type: string
          title: Ticker
        name:
          type: string
          title: Name
        class_code:
          type: string
          title: Class Code
        instrument_type:
          type: string
          title: Instrument Type
        category:
          type: string
          title: Category
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
        lot:
          type: integer
          title: Lot
        min_price_increment:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Min Price Increment
        api_trade_available:
          type: boolean
          title: Api Trade Available
        is_active:
          type: boolean
          title: Is Active
        is_selected:
          type: boolean
          title: Is Selected
        first_seen_at:
          type: string
          format: date-time
          title: First Seen At
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
        unit_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Unit Price
        lot_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Lot Price
        price_captured_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Price Captured At
      additionalProperties: false
      type: object
      required:
      - id
      - broker_id
      - instrument_id
      - figi
      - ticker
      - name
      - class_code
      - instrument_type
      - category
      - currency
      - lot
      - min_price_increment
      - api_trade_available
      - is_active
      - is_selected
      - first_seen_at
      - last_seen_at
      - unit_price
      - lot_price
      - price_captured_at
      title: CatalogInstrumentPriceSchema
    CatalogInstrumentSchema:
      properties:
        id:
          type: string
          title: Id
        broker_id:
          type: string
          title: Broker Id
        instrument_id:
          type: string
          title: Instrument Id
        figi:
          type: string
          title: Figi
        ticker:
          type: string
          title: Ticker
        name:
          type: string
          title: Name
        class_code:
          type: string
          title: Class Code
        instrument_type:
          type: string
          title: Instrument Type
        category:
          type: string
          title: Category
        currency:
          anyOf:
          - type: string
          - type: 'null'
          title: Currency
        lot:
          type: integer
          title: Lot
        min_price_increment:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Min Price Increment
        api_trade_available:
          type: boolean
          title: Api Trade Available
        is_active:
          type: boolean
          title: Is Active
        is_selected:
          type: boolean
          title: Is Selected
        first_seen_at:
          type: string
          format: date-time
          title: First Seen At
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
      additionalProperties: false
      type: object
      required:
      - id
      - broker_id
      - instrument_id
      - figi
      - ticker
      - name
      - class_code
      - instrument_type
      - category
      - currency
      - lot
      - min_price_increment
      - api_trade_available
      - is_active
      - is_selected
      - first_seen_at
      - last_seen_at
      title: CatalogInstrumentSchema
    CatalogListResponseSchema:
      properties:
        broker_id:
          type: string
          title: Broker Id
        items:
          items:
            $ref: '#/components/schemas/CatalogInstrumentPriceSchema'
          type: array
          title: Items
        categories:
          items:
            $ref: '#/components/schemas/CatalogCategorySchema'
          type: array
          title: Categories
        sync_state:
          $ref: '#/components/schemas/CatalogSyncStateSchema'
        currencies:
          items:
            type: string
          type: array
          title: Currencies
      additionalProperties: false
      type: object
      required:
      - broker_id
      - items
      - categories
      - sync_state
      - currencies
      title: CatalogListResponseSchema
...
    CreateAutomationSchema:
      properties:
        account_id:
          type: string
          title: Account Id
        strategy:
          anyOf:
          - $ref: '#/components/schemas/StrategySchema-Input'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - account_id
      title: CreateAutomationSchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InstrumentDetailsSchema:
      properties:
        broker_name:
          type: string
          title: Broker Name
        instrument:
          $ref: '#/components/schemas/CatalogInstrumentSchema'
        last_price:
          anyOf:
          - $ref: '#/components/schemas/LastPriceSchema'
          - type: 'null'
        lot_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Lot Price
        sync_state:
          $ref: '#/components/schemas/CatalogSyncStateSchema'
      additionalProperties: false
      type: object
      required:
      - broker_name
      - instrument
      - last_price
      - lot_price
      - sync_state
      title: InstrumentDetailsSchema
...
    InstrumentSelectionSchema:
      properties:
        selected:
          type: boolean
          title: Selected
      additionalProperties: false
      type: object
      required:
      - selected
      title: InstrumentSelectionSchema
...
    ReconciliationResultSchema:
      properties:
        broker_id:
          type: string
          title: Broker Id
        added:
          type: integer
          title: Added
        updated:
          type: integer
          title: Updated
        deactivated:
          type: integer
          title: Deactivated
        synchronized_at:
          type: string
          format: date-time
          title: Synchronized At
      additionalProperties: false
      type: object
      required:
      - broker_id
      - added
      - updated
      - deactivated
      - synchronized_at
      title: ReconciliationResultSchema
...
    TradingAutomationDetailsSchema:
      properties:
        automation:
          $ref: '#/components/schemas/AutomationSchema'
        operations:
          items:
            $ref: '#/components/schemas/OperationSchema'
          type: array
          title: Operations
        candles:
          items:
            $ref: '#/components/schemas/HistoricCandleSchema'
          type: array
          title: Candles
        errors:
          items:
            $ref: '#/components/schemas/PositionDetailsErrorSchema'
          type: array
          title: Errors
      additionalProperties: false
      type: object
      required:
      - automation
      - operations
      - candles
      - errors
      title: TradingAutomationDetailsSchema
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    moex_sentinel__views__brokers__BrokerConnectionSchema:
      properties:
        broker_id:
          type: string
          title: Broker Id
        available:
          type: boolean
          title: Available
        accounts_count:
          type: integer
          title: Accounts Count
      additionalProperties: false
      type: object
      required:
      - broker_id
      - available
      - accounts_count
      title: BrokerConnectionSchema
...